refactor(2d-render-pipeline): opt code
